requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0.3
django-crispy-forms==1.14.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.3.5
flake8==4.0.1
future==0.18.2
inflection==0.5.1
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.1
pytest-django==4.5.2
pytest-factoryboy==2.1.0
pytest-faker==2.0.0
python-dateutil==2.8.2
pytz==2022.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
